Make app/Console/cake pwd insensitive
With the previous versions of cake - you'd install cake in your path and
therefore when executed the pwd was a variable.

With the creation of app/Console/cake this is not true any more, and
copying the existing use of `pwd` for app means there is an implicit
assumption that it will always and only be ran from the app dir. If it
is ran from elsewhere - it's as if the core lib/Cake/Console/cake
had been run (app dir is "missing" no boostrap, no app-plugins etc.)

With this change, all of these act exactly the same:

    cd ~/root/app
    Console/cake

    cd ~/root/app/Console
    ./cake

    cd ~/root
    app/Console/cake

    cd /anywhere
    ~/root/app/Console/cake